size_t portability improved

+++ src/screen.c
@@ -18,7 +18,7 @@ Screen* screen_create(int h, int w, int y, int x){
 
 void screen_fill(Screen *s, const char* c, int col){
   wattron(s->win, COLOR_PAIR(col));
-  for(int i = 0; i < s->h; i++)
+  for(size_t i = 0; i < s->h; i++)
     mvwhline(s->win, i, 0, *c, s->w);
 
   wattroff(s->win, COLOR_PAIR(col));
@@ -29,8 +29,8 @@ void screen_fill_randomly(Screen *s, int seed, const char *c, int col, int chanc
 
   wattron(s->win, COLOR_PAIR(col));
 
-  for(int i = 0; i < s->h; i++)
-    for(int j = 0; j < s->w; j++)
+  for(size_t i = 0; i < s->h; i++)
+    for(size_t j = 0; j < s->w; j++)
       if((rand() % 100) +1 < chance)
         mvwprintw(s->win, i, j, c);
 
@@ -48,14 +48,14 @@ void screen_cross(Screen *s, const char* c, int col, int y, int x){
 void screen_grid(Screen *s, const char* c, int col, int spacing, int y_offset, int x_offset){
   int max = s->h > s->w ? s->h : s->w;
   spacing = spacing+1;
-  for(int i = 0; i < max; i += spacing)
+  for(size_t i = 0; i < max; i += spacing)
     screen_cross(s, c, col, i+y_offset, i+x_offset);
 }
 
 void screen_dotted_v_line(Screen *s, const char* c, int col, int y, int spacing){
   spacing *= 2;
   wattron(s->win, COLOR_PAIR(col));
-  for(int i = 0; i < s->w; i+=spacing*2){
+  for(size_t i = 0; i < s->w; i+=spacing*2){
     mvwhline(s->win, y, i, *c, spacing);
   }
 
@@ -65,7 +65,7 @@ void screen_dotted_v_line(Screen *s, const char* c, int col, int y, int spacing)
 void screen_dotted_h_line(Screen *s, const char* c, int col, int x, int spacing){
   wattron(s->win, COLOR_PAIR(col));
 
-  for(int i = 0; i < s->w; i+=spacing*2){
+  for(size_t i = 0; i < s->w; i+=spacing*2){
     mvwvline(s->win, i, x+x, *c, spacing);
     mvwvline(s->win, i, x+x+1, *c, spacing);
   }
@@ -75,9 +75,9 @@ void screen_dotted_h_line(Screen *s, const char* c, int col, int x, int spacing)
 void screen_polka_dot(Screen *s, const char* c, int col, int spacing){
   wattron(s->win, COLOR_PAIR(col));
 
-  for(int i = 0; i < s->w; i += spacing*4)
-    for(int j = 0; j < spacing*2; j++)
-      for(int k = 0; k < s->h; k+=spacing*2)
+  for(size_t i = 0; i < s->w; i += spacing*4)
+    for(size_t j = 0; j < spacing*2; j++)
+      for(size_t k = 0; k < s->h; k+=spacing*2)
         mvwvline(s->win, k, j+i, *c, spacing);
 
   wattroff(s->win, COLOR_PAIR(col));
@@ -86,9 +86,9 @@ void screen_polka_dot(Screen *s, const char* c, int col, int spacing){
 void screen_checker(Screen *s, const char* c, int col, int spacing){
   wattron(s->win, COLOR_PAIR(col));
 
-  for(int i = 0; i < s->w; i+=spacing*4)
-    for(int j = 0; j < spacing*2; j++)
-      for(int k = 0; k < s->h; k += spacing * 2){
+  for(size_t i = 0; i < s->w; i+=spacing*4)
+    for(size_t j = 0; j < spacing*2; j++)
+      for(size_t k = 0; k < s->h; k += spacing * 2){
         mvwvline(s->win, k, j+i, *c, spacing);
         mvwvline(s->win, k+spacing, j+i+spacing*2, *c, spacing);
       }
